+++ R/explorePgtraj.R
@@ -1,25 +1,59 @@
-# library(sf)
-library(lubridate)
-library(shiny)
-library(leaflet)
-# library(dplyr)
-# library(DBI)
-library(htmltools)
-library(mapview)
-library(shinyWidgets)
-library(testthat)
-# raster handling
-library(rgdal)
-library(raster)
-
-source("./shiny/utils.R")
-source("./shiny/createShinyView.R")
-
-
-# Shiny App----------------------------------------------------------------
+# # library(sf)
+# library(lubridate)
+# library(shiny)
+# library(leaflet)
+# # library(dplyr)
+# # library(DBI)
+# library(htmltools)
+# library(mapview)
+# library(shinyWidgets)
+# library(testthat)
+# 
+# source("./shiny/utils.R")
+# source("./shiny/createShinyView.R")
+# 
+# 
+# # Shiny App----------------------------------------------------------------
 
 
-pgtrajPlotter <-
+#' Explore a pgtraj interactively in a Shiny app
+#'
+#' @param conn DBI::DBIConnection
+#' @param schema String. Schema name of the pgtraj.
+#' @param pgtraj String. Pgtraj name.
+#' @param layers_vector List of character vectors. As c(schema, table).
+#' @param layers_params_vector Named list of lists. Names need to map to the
+#'  table names in layers_vector. Sub-lists contain parameters passed to
+#'  leaflet::add*. See example.
+#' @param layer_raster raster::RasterLayer object
+#' @param layers_params_raster List. Parameters passed to leaflet::addRasterImage()
+#'
+#' @return nothing
+#' @export
+#' 
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#'
+#' @examples
+#' \dontrun{
+#' # Vectore base layers to include
+#' layers_vector <- list(c("example_data", "county_subdiv"),
+#'                       c("example_data", "test_points")
+#'                       )
+#' layers_params_vector <- list(test_points=list(color = "red",
+#'                                               stroke = FALSE,
+#'                                               fillOpacity = 0.5),
+#'                              county_subdiv=list(color = "grey",
+#'                                                 fillOpacity = 0.2)
+#'                              )
+#' 
+#' # Raster base layers to include
+#' ras <- rgdal::readGDAL("./temp_data/florida_dem_county099.tif")
+#' ras2 <- raster::raster(ras, 1)
+#' ras2_leaflet <- leaflet::projectRasterForLeaflet(ras2)
+#' explorePgtraj(conn, schema, pgtraj, layers_vector, layers_params_vector,
+#'               layer_raster=ras2_leaflet)
+#' }
+explorePgtraj <-
     function(conn,
              schema,
              pgtraj,
@@ -29,42 +63,37 @@ pgtrajPlotter <-
              layers_params_raster=NULL) {
         view <- paste0("step_geometry_shiny_", pgtraj)
         # Get default time parameters
-        time_params <- get_traj_defaults(conn, schema, view, pgtraj)
+        time_params <- getTrajDefaults(conn, schema, view, pgtraj)
         
         tzone <- time_params$time_zone
         
-        tstamp_last <- as.POSIXct(time_params$tstamp_last,
-                                  origin = "1970-01-01 00:00:00",
-                                  tz = "UTC")
-        attributes(t)$tzone <- tzone
-        
-        t <- as.POSIXct(time_params$tstamp_start,
-                        origin = "1970-01-01 00:00:00",
-                        tz = "UTC")
+        # tstamp_last <- as.POSIXct(time_params$tstamp_last,
+        #                           origin = "1970-01-01 00:00:00",
+        #                           tz = "UTC")
+        # 
+        # tstamp_start <- as.POSIXct(time_params$tstamp_start,
+        #                 origin = "1970-01-01 00:00:00",
+        #                 tz = "UTC")
         # R uses time zone abbreviation to print time stamps,
-        # thus get_step_window while pgtraj stores the "long" time zone
+        # and also getStepWindow. On the other hand, pgtraj stores the "long" time zone
         # format (e.g. America/New_York instead of EDT). Thus the warning
         # of In check_tzones(e1, e2) : 'tzone' attributes are inconsistent
-        attributes(t)$tzone <- tzone
+        # attributes(time_params$tstamp_start)$tzone <- tzone
         
-        increment <- period(num = time_params$increment,
+        increment <- lubridate::period(num = time_params$increment,
                               units = "seconds")
         
         # default interval is 10*increment (~10 steps)
-        limit <- t + (increment * 10)
-        if (limit < tstamp_last) {
+        limit <- time_params$tstamp_start + (increment * 10)
+        if (limit < time_params$tstamp_last) {
             interval <- increment * 10
         } else {
             message("Loading full trajectory, because it is shorter than 10 steps.")
-            interval <- tstamp_last - t
+            interval <- time_params$tstamp_last - time_params$tstamp_start
         }
         
-        # Get initial set of trajectories
-        st <-
-            get_step_window(conn, schema, view, t, interval, FALSE, info_cols)
-        
         # Get full traj
-        st_1 <- get_full_traj(conn, schema, view)
+        st_1 <- getFullTraj(conn, schema, view)
         
         # color by animal_name
         # factpal <- colorFactor(topo.colors(4), st$animal_name)
@@ -76,7 +105,7 @@ pgtrajPlotter <-
                                      na.color = "#808080")
         
         # get burst list for burst mode
-        bursts_df <- get_bursts_df(conn, schema, view)
+        bursts_df <- getBurstsDF(conn, schema, view)
         burst_len <- nrow(bursts_df)
         colors_burst <- colorFactor(topo.colors(burst_len),
                                     bursts_df$burst_name,
@@ -102,14 +131,7 @@ pgtrajPlotter <-
             raster_name <- NULL
         }
         
-        # infolocs columns
-        infolocs_table <- paste0("infolocs_", pgtraj)
-        ic <- getInfolocsColumns(conn, schema, infolocs_table)
-        if(nrow(ic) > 0) {
-            info_cols <- paste(paste(ic$column_name, collapse = ", "), ",")
-        } else {
-            info_cols <- NULL
-        }
+        info_cols <- getInfolocsColumns(conn, schema, pgtraj)
             
         ui <-
             fluidPage(# tags$style(type = "text/css", "html, body {width:100%;height:100%}"),
@@ -190,9 +212,9 @@ pgtrajPlotter <-
                     sliderInput(
                         "range",
                         "Time window:",
-                        min = t,
-                        max = tstamp_last,
-                        value = c(t, t + interval),
+                        min = time_params$tstamp_start,
+                        max = time_params$tstamp_last,
+                        value = c(time_params$tstamp_start, time_params$tstamp_start + interval),
                         step = increment,
                         timezone = tzone
                     ),
@@ -208,7 +230,7 @@ pgtrajPlotter <-
             w <- reactiveValues(data = st_1)
             x <-
                 reactiveValues(
-                    currStep = st,
+                    currStep = NULL,
                     counter = 0,
                     burst_counter = 0,
                     burst_name = NULL,
@@ -216,7 +238,7 @@ pgtrajPlotter <-
                 )
             
             # get current time window and the next
-            timeOut <- reactiveValues(currTime = t,
+            timeOut <- reactiveValues(currTime = time_params$tstamp_start,
                                       interval = interval,
                                       increment = increment,
                                       increment_unit = unit_init,
@@ -226,21 +248,26 @@ pgtrajPlotter <-
             observeEvent(input$step_mode, {
                 x$counter <- x$counter + 1
                 x$currStep <-
-                    get_step_window(conn,
-                                    schema,
-                                    view,
-                                    timeOut$currTime,
-                                    timeOut$interval,
-                                    input$step_mode,
-                                    info_cols)
-            })
+                    getStepWindow(
+                        conn,
+                        schema,
+                        view,
+                        timeOut$currTime,
+                        timeOut$interval,
+                        input$step_mode,
+                        info_cols,
+                        time_params$tstamp_start,
+                        time_params$tstamp_last
+                    )
+            },
+            ignoreInit = TRUE)
             
             # convert values in Increment to the selected unit
             observeEvent(input$increment_unit, {
                 if(is.null(input$increment) | is.logical(input$increment)){
                     return()
                 }
-                timeOut$increment <- as.period(timeOut$increment,
+                timeOut$increment <- lubridate::as.period(timeOut$increment,
                                                unit = input$increment_unit)
                 
                 updateNumericTimeInput(session, input$increment_unit,
@@ -252,7 +279,7 @@ pgtrajPlotter <-
                 if(is.null(input$interval) | is.logical(input$interval)){
                     return()
                 }
-                timeOut$interval <- as.period(timeOut$interval,
+                timeOut$interval <- lubridate::as.period(timeOut$interval,
                                               unit = input$interval_unit)
                 
                 updateNumericTimeInput(session, input$interval_unit,
@@ -261,102 +288,122 @@ pgtrajPlotter <-
             
             # set Increment from input field
             observeEvent(input$increment, {
-                if(is.null(input$increment) | is.logical(input$increment)){
+                # do not update Increment in case the input is 0 or empty
+                if (is.null(input$increment) |
+                    is.logical(input$increment) |
+                    identical(input$increment, as.integer(0))) {
                     return()
                 }
-                timeOut$increment <- setTimeInput(input$increment_unit,
-                                                 input$increment,
-                                                    timeOut$increment)
+                timeOut$increment <-
+                    setTimeInput(input$increment_unit,
+                                 input$increment,
+                                 timeOut$increment)
             })
             
             # set Interval from input field
             observeEvent(input$interval, {
-                if(is.null(input$interval) | is.logical(input$interval)){
+                # do not update Interval in case the input is 0 or empty
+                if (is.null(input$interval) |
+                    is.logical(input$interval) |
+                    identical(input$interval, as.integer(0))) {
                     return()
+                } else {
+                    timeOut$interval <-
+                        setTimeInput(input$interval_unit,
+                                     input$interval,
+                                     timeOut$interval)
+                    
+                    # update time window slider
+                    if (period_to_seconds(timeOut$interval) > lubridate::period(0)) {
+                        updateSliderInput(
+                            session,
+                            "range",
+                            value = c(
+                                timeOut$currTime,
+                                timeOut$currTime + timeOut$interval
+                            ),
+                            step = timeOut$increment
+                        )
+                    }
                 }
-                timeOut$interval <- setTimeInput(input$interval_unit, input$interval,
-                             timeOut$interval)
-                
-                # update time window slider
-                if(period_to_seconds(timeOut$interval) > period(0)) {
-                    updateSliderInput(
-                        session,
-                        "range",
-                        value = c(timeOut$currTime,
-                                  timeOut$currTime + timeOut$interval),
-                        step = timeOut$increment
-                    )
-                }
-            
             })
             
             # set Interval and Time Window from slider
             observeEvent(input$range, {
-                timeOut$currTime <- input$range[1]
+                # the time window must be "open" in order to increment the time stamp
+                # do nothing if the start and end times are equal
+                stime <- input$range[1]
+                etime <- input$range[2]
                 
-                timeOut$interval <-
-                    as.period(input$range[2] - input$range[1])
-                x$counter <- x$counter + 1
-                x$currStep <-
-                    get_step_window(conn,
-                                    schema,
-                                    view,
-                                    timeOut$currTime,
-                                    timeOut$interval,
-                                    input$step_mode,
-                                    info_cols)
-                
-                # update the Interval numeric input
-                # updateNumericTimeInput(session, input$interval_unit,
-                #                        "interval", timeOut$interval)
+                if(stime < etime) {
+                    timeOut$currTime <- stime
+                    
+                    timeOut$interval <- lubridate::as.period(etime - stime)
+                    # for assigning alternating group names in order to 
+                    # remove the previous step from the plot
+                    x$counter <- x$counter + 1
+                    
+                    x$currStep <-
+                        getStepWindow(conn,
+                                        schema,
+                                        view,
+                                        timeOut$currTime,
+                                        timeOut$interval,
+                                        input$step_mode,
+                                        info_cols,
+                                        time_params$tstamp_start,
+                                        time_params$tstamp_last)
+                    
+                    # update the Interval numeric input
+                    # updateNumericTimeInput(session, input$interval_unit,
+                    #                        "interval", timeOut$interval)
+                } else {
+                    # ignore input
+                }
             })
             
             # Only update timestamp on click
             observeEvent(input$n, {
-                # for assigning alternating group names
-                x$counter <- x$counter + 1
+                # the time window must be "open" in order to increment the time stamp
+                # do nothing if the start and end times are equal
+                stime <- timeOut$currTime + timeOut$increment
+                etime <- stime + timeOut$interval
                 
-                # timeOut$currTime <- timeOut$currTime + timeOut$increment
-                nt <- timeOut$currTime + timeOut$increment
-                
-                # update time window slider
-                updateSliderInput(
-                    session,
-                    "range",
-                    value = c(
-                        nt,
-                        nt + timeOut$interval
-                    ),
-                    step = timeOut$increment
-                )
+                if(stime < etime) {
+                    # update time window slider
+                    updateSliderInput(
+                        session,
+                        "range",
+                        value = c(
+                            stime,
+                            etime
+                        ),
+                        step = timeOut$increment
+                    )
+                } else {
+                    message("time window out of range")
+                }
             })
             
             observeEvent(input$b, {
-                # for assigning alternating group names
-                x$counter <- x$counter + 1
-
-                nt <- timeOut$currTime - timeOut$increment
+                stime <- timeOut$currTime - timeOut$increment
+                etime <- stime + timeOut$interval
                 
-                # update time window slider
-                updateSliderInput(
-                    session,
-                    "range",
-                    value = c(
-                        nt,
-                        nt + timeOut$interval
-                    ),
-                    step = timeOut$increment
-                )
+                if(stime < etime) {
+                    # update time window slider
+                    updateSliderInput(
+                        session,
+                        "range",
+                        value = c(
+                            stime,
+                            etime
+                        ),
+                        step = timeOut$increment
+                    )
+                } else {
+                    message("time window out of range")
+                }
             })
-            
-            # # Report current timestamp
-            # output$tstamp <- renderText({
-            #     paste(
-            #         format(timeOut$currTime, usetz = TRUE),
-            #         "-",
-            #         format(timeOut$currTime + timeOut$interval, usetz = TRUE)
-            #     )
-            # })
             
             output$map <- renderLeaflet({
                 if (is.null(w$data)) {
@@ -464,39 +511,49 @@ pgtrajPlotter <-
             })
             
             observe({
-                # counter for adding/removing groups
-                if (x$counter %% 2 == 0) {
-                    gname <- "traj"
-                } else {
-                    gname <- "trajnew"
-                }
-                # colors
-                if(input$color_choice == "Bursts"){
-                    colorpal <- ~colors_burst(burst_name)
-                } else {
-                    colorpal <- ~colors_animal(animal_name)
-                }
-                # map
-                proxy <- leafletProxy("map") %>%
-                    addPolylines(
-                        data = x$currStep,
-                        group = gname,
-                        fillOpacity = 1,
-                        opacity = 1,
-                        color = colorpal,
-                        weight = 4,
-                        popup = mapview::popupTable(x$currStep)
-                    )
-                if (x$counter %% 2 == 0) {
-                    proxy %>% clearGroup("trajnew")
-                } else {
-                    proxy %>% clearGroup("traj")
-                }
-                
-                # because observeEven doesn't pass value when all burst are
-                # deselected
-                if (is.null(input$burst_picker)) {
-                    proxy %>% clearGroup("bursts")
+                # don't do anything when there is no geometry to display
+                if(!is.null(x$currStep)) {
+                    # counter for adding/removing the next/previous set of steps
+                    # when plotting a trajectory
+                    if (x$counter %% 2 == 0) {
+                        gname <- "traj"
+                    } else {
+                        gname <- "trajnew"
+                    }
+                    # colors
+                    if(input$color_choice == "Bursts"){
+                        colorpal <- ~colors_burst(burst_name)
+                    } else {
+                        colorpal <- ~colors_animal(animal_name)
+                    }
+                    # map
+                    if(length(st_geometry(x$currStep)) > 0) {
+                        proxy <- leafletProxy("map") %>%
+                            addPolylines(
+                                data = x$currStep,
+                                group = gname,
+                                fillOpacity = 1,
+                                opacity = 1,
+                                color = colorpal,
+                                weight = 4,
+                                popup = mapview::popupTable(x$currStep)
+                            )
+                        
+                        if (x$counter %% 2 == 0) {
+                            proxy %>% clearGroup("trajnew")
+                        } else {
+                            proxy %>% clearGroup("traj")
+                        }
+                        
+                        # because observeEven doesn't pass value when all burst are
+                        # deselected
+                        if (is.null(input$burst_picker)) {
+                            proxy %>% clearGroup("bursts")
+                        }
+                        
+                    } else {
+                        # leaflet crashes
+                    }
                 }
             })