+++ shiny/utils_shiny.R
@@ -3,17 +3,36 @@ library(lubridate)
 library(dplyr)
 library(DBI)
 
-# Queries ------------------------------------------------------------
-
-# Get steps within a temporal window
-get_step_window <- function(conn, schema, view, time, interval, step_mode,
-                            info_cols){
+#' Get steps within a temporal window
+#'
+#' @param conn DBI::DBIConnection
+#' @param schema String. Schema name.
+#' @param view String. View name.
+#' @param time String of the start time of the time window. Including time zone.
+#' @param interval lubridate::Period object of the time window
+#' @param step_mode Boolean. Detailed step info (TRUE) or aggregate
+#' @param info_cols Character vector of the infolocs columns of the pgtraj.
+#' @param tstamp_start POSIXct with timestamp. First time stamp in view.
+#' @param tstamp_last POSIXct with timestamp. Last time stamp in view.
+#'
+#' @return A simple feature object of the steps. NULL when out of range.
+#' 
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
+getStepWindow <- function(conn, schema, view, time, interval, step_mode,
+                            info_cols, tstamp_start, tstamp_last){
     stopifnot(is.period(interval))
     i <- period_to_seconds(interval)
     t <- dbQuoteString(conn, format(time, usetz = TRUE))
     t_interval <- dbQuoteString(conn, paste(i, "seconds"))
     schema_q <- dbQuoteIdentifier(conn, schema)
     view_q <- dbQuoteIdentifier(conn, view)
+    
+    # if((time < tstamp_start | time > tstamp_last) |
+    #    (i < 1)) {
+    #     message("time window out of range")
+    #     return(NULL)
+    # }
     
     if(step_mode){
         sql_query <- paste0("
@@ -53,13 +72,28 @@ get_step_window <- function(conn, schema, view, time, interval, step_mode,
                             t_interval, "::INTERVAL)
                             GROUP BY
                                 burst_name, animal_name;")
-        
     }
-    return(st_read_db(conn, query=sql_query, geom_column = "step_geom"))
-    }
+    withCallingHandlers(
+        s <- st_read_db(conn, query = sql_query, geom_column = "step_geom"),
+        warning = function(w) {
+            warning(paste("Didn't find any steps at", time, "+", interval))
+        }
+    )
+    
+    return(s)
+}
 
-# Get list of bursts in step_geometry view
-get_bursts_df <- function(conn, schema, view){
+#' Get distinct burst names in a step_geometry view
+#'
+#' @param conn DBI::DBIConnection
+#' @param schema String. Schema name.
+#' @param view String. View name.
+#'
+#' @return data frame with column 'burst_name'
+#' 
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
+getBurstsDF <- function(conn, schema, view){
     schema_q <- dbQuoteIdentifier(conn, schema)
     view_q <- dbQuoteIdentifier(conn, view)
     sql_query <- paste0("
@@ -70,7 +104,16 @@ get_bursts_df <- function(conn, schema, view){
     return(dbGetQuery(conn, sql_query))
 }
 
-# Get list of animals in step_geometry view
+#' Get distinct animal names in step_geometry view
+#'
+#' @param conn DBI::DBIConnection
+#' @param schema String. Schema name.
+#' @param view String. View name.
+#'
+#' @return data frame with column 'animal_name'
+#'
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
 getAnimalsDf <- function(conn, schema, view){
     schema_q <- dbQuoteIdentifier(conn, schema)
     view_q <- dbQuoteIdentifier(conn, view)
@@ -82,9 +125,18 @@ getAnimalsDf <- function(conn, schema, view){
     return(dbGetQuery(conn, sql_query))
 }
 
-# Get geometry of bursts as linestring
+#' Get geometry of bursts as linestring
+#'
+#' @param conn DBI::DBIConnection
+#' @param schema String. Schema name.
+#' @param view String. View name.
+#' @param burst_name String. Accepts a character vector of variable length
+#'
+#' @return a single LINESTRING per burst
+#'
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
 getBurstGeom <- function(conn, schema, view, burst_name){
-    # accepts a character vector of variable length
     
     if (is.null(burst_name) | length(burst_name) == 0){
         return()
@@ -106,8 +158,17 @@ getBurstGeom <- function(conn, schema, view, burst_name){
     return(st_read_db(conn, query=sql_query, geom_column = "burst_geom"))
 }
 
-# Get the complete trajectory of an animal as a single linestring
-get_full_traj <- function(conn, schema, view){
+#' Get the complete trajectory of an animal as a single linestring
+#'
+#' @param conn DBI::DBIConnection
+#' @param schema String. Schema name.
+#' @param view String. View name.
+#'
+#' @return a single LINESTRING per animal
+#'
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
+getFullTraj <- function(conn, schema, view){
     sql_query <- paste0("
                         SELECT
                         st_makeline(step_geom)::geometry(linestring, 4326) AS traj_geom,
@@ -117,8 +178,18 @@ get_full_traj <- function(conn, schema, view){
     return(st_read_db(conn, query=sql_query, geom_column = "traj_geom"))
 }
 
-# Get default time parameters
-get_traj_defaults <- function(conn, schema, view, pgtraj){
+#' Get default time parameters for steps
+#'
+#' @param conn DBI::DBIConnection
+#' @param schema String. Schema name.
+#' @param view String. View name.
+#' @param pgtraj String. Pgtraj name
+#'
+#' @return data frame with columns: tstamp_start (epoch), tstamp_last (epoch), increment, tzone
+#' 
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
+getTrajDefaults <- function(conn, schema, view, pgtraj){
     schema_q <- dbQuoteIdentifier(conn, schema)
     view_q <- dbQuoteIdentifier(conn, view)
     sql_query <- paste0("
@@ -131,16 +202,8 @@ get_traj_defaults <- function(conn, schema, view, pgtraj){
     # default increment is the median step duration
     sql_query <- paste0("
                         SELECT
-                            EXTRACT(
-                                epoch
-                            FROM
-                                MIN( DATE )
-                            ) AS tstamp_start,
-                            EXTRACT(
-                                epoch
-                            FROM
-                                MAX( DATE )
-                            ) AS tstamp_last,
+                            MIN( DATE ) AS tstamp_start,
+                            MAX( DATE ) AS tstamp_last,
                             PERCENTILE_CONT( 0.5 ) WITHIN GROUP(
                             ORDER BY
                                 dt
@@ -152,6 +215,17 @@ get_traj_defaults <- function(conn, schema, view, pgtraj){
     return(cbind(time_params, tzone))
 }
 
+#' Convert the value of input$interval/increment to the unit selected in input$*_unit
+#'
+#' @param session Shiny session
+#' @param inputUnit String. One of years, months, days, hours, minutes seconds
+#' @param inputId String. Id of the input slot.
+#' @param reactiveTime A lubridate::Period object stored in a Reactive Value
+#'
+#' @return nothing
+#'
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
 updateNumericTimeInput <- function(session, inputUnit, inputId, reactiveTime){
     if (inputUnit == "years") {
         updateNumericInput(session, inputId,
@@ -174,6 +248,16 @@ updateNumericTimeInput <- function(session, inputUnit, inputId, reactiveTime){
     }
 }
 
+#' Set a lubridate::Period value from input$interval/increment
+#'
+#' @param inputUnit String.
+#' @param inputTime lubridate::Period
+#' @param reactiveTime Reactive value to set
+#'
+#' @return nothing
+#'
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
 setTimeInput <- function(inputUnit, inputTime, reactiveTime){
     if (inputUnit == "years") {
         reactiveTime <- period(num = inputTime,
@@ -198,8 +282,22 @@ setTimeInput <- function(inputUnit, inputTime, reactiveTime){
     return(reactiveTime)
 }
 
-# layers <- list(c("example_data", "county_subdiv"), c("public", "florida_dem"))
-# return: list(name=sf object, name2=sf object)
+#' Get base layers from database
+#' 
+#' Not implemented for rasters
+#'
+#' @param conn DBI::DBIConnection
+#' @param layers List. List of character vectors for each layer to include as a 
+#' base layer. 
+#'
+#' @return list of simple features as \code{list(name=sf object, name2=sf object)}
+#'
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @examples 
+#' \dontrun{
+#' layers <- list(c("schema1", "tableA"), c("schema2", "tableB"))
+#' }
+#' @keywords internal
 getLayers <- function(conn, layers) {
     geo_type <- findGeoType(conn, layers)
     base <- list()
@@ -221,21 +319,27 @@ getLayers <- function(conn, layers) {
             warning("raster layers not implemented yet")
         }
     }
-    
     return(base)
 }
 
-# ras <- readGDAL(dsn) # Get your file as SpatialGridDataFrame
-# ras2 <- raster(ras,1) # Convert the first Band to Raster
-# plot(ras2)
-# 
-# rast <- pgGetRast(conn, c("public", "florida_dem"))
-
-# layers <- list(c("example_data", "county_subdiv"))
-# b <- getLayers(conn, layers)
-
-# layers <- list(c("example_data", "county_subdiv"), c("public", "florida_dem"))
-# geo_type$vect[[1]]
+#' Figures out whether the provided database relation contains vector or raster data.
+#'
+#' Not implemented for rasters.
+#'
+#' @param conn DBI::DBIConnection
+#' @param layers List. List of character vectors for each layer to include as a 
+#' base layer. 
+#'
+#' @return List of lists of database relations as \code{list(vect = list(), rast = list())}.
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#'
+#' @examples
+#' \dontrun{
+#' layers <- list(c("example_data", "county_subdiv"), c("public", "florida_dem"))
+#' geo_type <- findGeoType(conn, layers)
+#' geo_type$vect[[1]]
+#' }
+#' @keywords internal
 findGeoType <- function(conn, layers) {
     expect_true((length(layers) >= 1))
     # geo_type <- data.frame(name = character(), type = character(),
@@ -256,8 +360,15 @@ findGeoType <- function(conn, layers) {
     return(geo_type)
 }
 
-
-# layer: c(schema, table)
+#' Does a table contain vector data?
+#'
+#' @param conn DBI::DBIConnection
+#' @param layer String. As c(schema, table)
+#'
+#' @return Boolean
+#'
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
 isVector <- function(conn, layer) {
     sql_query <- paste0("SELECT *
                         FROM public.geometry_columns
@@ -273,7 +384,15 @@ isVector <- function(conn, layer) {
 }
 
 
-# layer: c(schema, table)
+#' Does a table contain raster data?
+#'
+#' @param conn DBI::DBIConnection
+#' @param layer String. As c(schema, table)
+#'
+#' @return Boolean
+#'
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
 isRaster <- function(conn, layer) {
     sql_query <- paste0("SELECT *
                         FROM public.raster_columns
@@ -287,3 +406,38 @@ isRaster <- function(conn, layer) {
         return(FALSE)
     }
 }
+
+
+#' Get all columns in the infolocs table but the step_id
+#' 
+#' Gets all the columns names in the infoloc table of the pgtraj and parses
+#' them for inserting into an SQL query, e.g.: "col1 ,col2, col2 ,"
+#'
+#' @param conn DBI::DBIConnection
+#' @param schema String. Schema name.
+#' @param pgtraj String. Pgtraj name.
+#'
+#' @return character vector or NULL if there are no infolocs
+#' 
+#' @author Balázs Dukai \email{balazs.dukai@@gmail.com}
+#' @keywords internal
+getInfolocsColumns <- function(conn, schema, pgtraj){
+    schema_s <- dbQuoteString(conn, schema)
+    table_s <- dbQuoteString(conn, paste0("infolocs_", pgtraj))
+    
+    sql_query <- paste0("
+                        SELECT column_name
+                        FROM information_schema.columns
+                        WHERE table_schema = ",schema_s,"
+                        AND table_name = ",table_s,"
+                        AND column_name != 'step_id';")
+    ic <- dbGetQuery(conn, sql_query)
+    
+    if(nrow(ic) > 0) {
+        info_cols <- paste(paste(ic$column_name, collapse = ", "), ",")
+    } else {
+        info_cols <- NULL
+    }
+    return(info_cols)
+}
+